# [IOTDB-6120] Push down limit/offset in query with group by time