desgin doc: https://apache-iotdb.feishu.cn/docx/TAsXd5hdZoVevrxqqgtcg9A9nhf

This PR optimized the performance of query by pushing down limit/offset when there is`group by time` clause in SQL.

We have created 10 devices and inserted 100,000 records for each device with timestamps incrementing from 0, 1, 2, and so on, to test performance. The execution times for different SQL queries before and after optimization are recorded in the following table:
<byte-sheet-html-origin data-version="4" data-is-embed="true" data-grid-line-hidden="false">

sql | original | optimized | orginal(dop=1) | optimized(dop=1) | explain
-- | -- | -- | -- | --| --|
select count(*) from root.sg.d0 group by ((0,100000],100ms) offset 100 limit 10 | 10.29 | 3.06 | 26.86 |  5.19 | Transfer group by time to ((10000,11000],100ms)
select count(*) from root.sg.d0 group by ((0,100000],100ms) order by time limit 10 offset 100 | 10.48 | 3.11 | 26.82 | 5.34 | The same as above
select count(*) from root.sg.d0 group by ((0,100000],100ms) order by time desc limit 10 offset 100 | 10.50 | 3.36 | 28.61 | 6.13 | Transfer group by time to ((89000,90000],100ms)
select count(*) from root.sg.d0 group by ((0,100000],100ms) offset 2000 limit 10 | 10.07| 3.49 | 27.21 | 4.39 | The `offset` skipped all the data, resulting in an empty result set.
select count(*) from root.sg.d0 group by ((0,100000],100ms) order by count(num) limit 10 offset 100 | 11.59 | 11.07 | 27.13 | 27.14 | Using `order by` clause other than "order by time" does not allow for query optimization through query pushdown, so the normal query process is followed |
select count(*) from root.sg.d0 group by ((0,100000],100ms) order by count(num) desc limit 10 offset 100 | 10.44 | 11.03 | 27.41 | 27.58 | The same as above
select count(*) from root.** group by ((0,100000],100ms) offset 1100 limit 4200 align by device | 48.13 | 37.35|144.57|122.11|Optimized 10 devices to device1 to device5.
select count(*) from root.** group by ((0,100000],100ms) offset 1100 limit 200 align by device | 23.51 | 6.38 |43.11|9.75| Optimized to group by time((10000, 30000], 100ms) for device1, while getting rid of the other 9 devices.
select count(*) from root.** group by ((0,100000],100ms) offset 1100 limit 20 align by device | 18.38 | 4.57 |42.77|5.74|Optimized to group by time((10000, 12000], 100ms) for device1, while getting rid of the other 9 devices.
select count(*) from root.** group by ((0,100000],100ms) offset 100 align by device | 57.69 | 59.57|246.58|246.47| The result set includes data from all devices.
</byte-sheet-html-origin>

From the performance test results, it can be observed that optimized code for queries with time parameter pushdown achieves better performance. Additionally, even in cases where parameter pushdown is not feasible, stable performance is ensured. In the context of "align by device," discarding devices also leads to objective performance enhancement. When no devices are discarded, the original performance can still be maintained.